[backend] feat/hexagonal-create-deck: restructure app into hexagonal architecture with DI, domain exceptions and global handlers
