+++ backend/app/api/__init__.py
@@ -1,4 +1,4 @@
-from .controllers.deck import deck_router
-from .controllers.health import health_router
+from .controllers.deck_controller import deck_router
+from .controllers.health_controller import health_router
 
 __all__ = ['deck_router', 'health_router']
