+++ backend/app/api/controllers/health_controller.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
 
-from app.api.schemas.health import HealthOut
+from app.api.schemas.health_schema import HealthOut
 
 health_router = APIRouter(tags=['Infra'])
