+++ backend/app/core/repositories/deck_repository.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from typing import Optional
 from uuid import UUID
 
 from app.domain.entities.deck import Deck
@@ -9,7 +10,7 @@ class IDeckRepository(ABC):
     async def find_all(self) -> list[Deck]: ...
 
     @abstractmethod
-    async def find_by_id(self, deck_id: UUID) -> Deck | None: ...
+    async def find_by_id(self, deck_id: UUID) -> Optional[Deck]: ...
 
     @abstractmethod
     async def create(self, deck: Deck) -> Deck: ...
