+++ backend/app/domain/exceptions/__init__.py
@@ -1,0 +1,3 @@
+from app.domain.exceptions.deck_exceptions import DeckValidationError
+
+__all__ = ['DeckValidationError']
