+++ backend/app/domain/exceptions/deck_exceptions.py
@@ -1,0 +1,2 @@
+class DeckValidationError(ValueError):
+    pass
